Refactor starknet_sierra_compile.
