+++ crates/starknet/src/lib.rs
@@ -1,2 +1,3 @@
 pub mod abi;
+pub mod casm_contract_class;
 pub mod contract_class;
